reafctor: modificão no código

+++ src/main/resources/db/migration/V1__create_initial_tables.sql
@@ -1,5 +1,5 @@
 CREATE TABLE stadiums (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     city VARCHAR(255) NOT NULL,
     capacity INTEGER,
@@ -7,20 +7,20 @@ CREATE TABLE stadiums (
 );
 
 CREATE TABLE clubs (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     founded DATE NOT NULL,
     url_img VARCHAR(255),
-    stadium_id INTEGER,
+    stadium_id BIGINT,
     CONSTRAINT fk_stadium_club FOREIGN KEY (stadium_id) REFERENCES stadiums(id)
 );
 
 CREATE TABLE players (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     position VARCHAR(255),
     shirt_number INTEGER,
     url_img VARCHAR(255),
-    club_id INTEGER,
+    club_id BIGINT,
     CONSTRAINT fk_club_player FOREIGN KEY (club_id) REFERENCES clubs(id)
 );
